Change screenshot naming, remove (trying to) overriding vanilla screenshot key.
No point trying to override the vanilla screenshot key. Better to just
configure the custom screenshot key and the vanilla screenshot key
in the controls menu. It didn't even work anyway.

Changed the screenshot name format slightly, moved the number increment
to the end instead of the middle.
Also changed the cryptic "_c" into "_scc" as in sourcecodedcore.